fix: Enable janitor scheduler in HTTP mode with sub-hour intervals
- HTTP mode is ideal for scheduled tasks (persistent daemon process)
- stdio mode is ephemeral (dies when client disconnects)
- Support fractional interval_hours (e.g., 0.5 for 30 minutes)
- Use admin_token for scheduled runs in HTTP mode
- Remove incorrect transport mode restriction
- Production config set to run every 30 minutes

The original code had backwards logic - HTTP servers are long-running
daemons perfect for background jobs, while stdio processes are tied
to client sessions and terminate when clients disconnect.

+++ internal/config/config.go
@@ -59,7 +59,7 @@ type LogConfig struct {
 // JanitorConfig holds janitor scheduling configuration.
 type JanitorConfig struct {
 	Enabled               bool     `toml:"enabled"`                // Enable scheduled janitor runs
-	IntervalHours         int      `toml:"interval_hours"`         // How often to run (in hours)
+	IntervalHours         float64  `toml:"interval_hours"`         // How often to run (in hours, supports fractional values like 0.5)
 	CreateProposal        bool     `toml:"create_proposal"`        // Auto-create proposals for critical issues
 	CreateImprovements    bool     `toml:"create_improvements"`    // Auto-create Improvement entities from findings
 	ImprovementThresholds []string `toml:"improvement_thresholds"` // Severity levels that trigger improvements: ["critical", "warning"]
@@ -102,7 +102,7 @@ func Load(configPath string) (*Config, error) {
 		},
 		Janitor: JanitorConfig{
 			Enabled:               false,                           // Disabled by default
-			IntervalHours:         1,                               // Run every hour when enabled
+			IntervalHours:         1.0,                             // Run every hour when enabled
 			CreateProposal:        false,                           // Don't auto-create proposals by default
 			CreateImprovements:    false,                           // Don't auto-create improvements by default
 			ImprovementThresholds: []string{"critical", "warning"}, // When enabled, create improvements for critical and warning issues
@@ -226,8 +226,8 @@ func (c *Config) applyEnv() {
 		c.Janitor.Enabled = (v == "true" || v == "1")
 	}
 	if v := os.Getenv("SPECMCP_JANITOR_INTERVAL_HOURS"); v != "" {
-		var hours int
-		if _, err := fmt.Sscanf(v, "%d", &hours); err == nil && hours > 0 {
+		var hours float64
+		if _, err := fmt.Sscanf(v, "%f", &hours); err == nil && hours > 0 {
 			c.Janitor.IntervalHours = hours
 		}
 	}
